new advancement

+++ app/src/main/java/com/example/glucoguardclient/data/send/PostData.kt
@@ -2,12 +2,12 @@ package com.example.glucoguardclient.data.send
 
 data class PostData(
     val pregnancies: Int,
-    val Glucose: Int,
-    val BloodPressure: Int,
-    val SkinThickness: Int,
-    val Insulin: Int,
-    val BMI: Int,
-    val DiabetesPedigreeFunction: Int,
+    val Glucose: Double,
+    val BloodPressure: Double,
+    val SkinThickness: Double,
+    val Insulin: Double,
+    val BMI: Double,
+    val DiabetesPedigreeFunction: Double,
     val Age: Int
 ){
